+++ core/src/towers/basic_tower/mod.rs
@@ -1,5 +1,7 @@
 mod components;
+mod systems;
 mod utils;
 
 pub use self::components::*;
+pub use self::systems::*;
 pub use self::utils::*;
