+++ core/src/towers/basic_tower/utils.rs
@@ -1,8 +1,5 @@
+use super::super::components::{TowerMarker, TowerPosition};
 use bevy::{ecs::system::SystemState, prelude::*};
-
-use crate::towers::components::{
-    TowerMarker, TowerModel, TowerPosition,
-};
 
 pub fn spawn_model(
     commands: &mut Commands,
@@ -35,25 +32,4 @@ pub fn spawn(world: &mut World, pos: Vec3) {
     ));
 
     state.apply(world);
-}
-
-/**
- * Render models for newly-created towers
- */
-pub fn render(
-    mut commands: Commands,
-    mut meshes: ResMut<Assets<Mesh>>,
-    mut materials: ResMut<Assets<StandardMaterial>>,
-    towers: Query<(Entity, &TowerPosition), Without<TowerModel>>,
-) {
-    for (entity, pos) in towers.iter() {
-        let model = spawn_model(
-            &mut commands,
-            &mut meshes,
-            &mut materials,
-            Vec3::new(pos.x as f32, 0.0, pos.z as f32),
-        );
-
-        commands.entity(entity).insert(TowerModel(model));
-    }
 }
