+++ core/src/units/basic_unit/utils.rs
@@ -1,12 +1,9 @@
-use bevy::prelude::*;
-
-use crate::{
-    gui::console,
-    units::components::{
-        UnitDist, UnitHealth, UnitMarker, UnitPathId, UnitSpawnTick,
-        UnitStatus, UnitStatusTypes,
-    },
+use super::super::components::{
+    UnitDist, UnitHealth, UnitMarker, UnitPathId, UnitSpawnTick,
+    UnitStatus, UnitStatusTypes,
 };
+use crate::gui::console;
+use bevy::prelude::*;
 
 pub fn spawn(commands: &mut Commands, id_path: u8, tick: u32) {
     commands.spawn((
